Prevent potentially applying an index twice
In Clusters.Add when accessing c.indexers there was no locking applied,
that could produce a concurrent access error as well as loosing an
index.

E.g. If a cluster is added and while the clusters is engaged with the
Aware another index is added - then the index would be added on the
cluster in the .IndexField and then again when .Add iterates over the
indizes.

Instead .IndexField copies .clusters while holding the lock and .Add
read locks to synchronize these two methods.

Signed-off-by: Nelo-T. Wallus <red.brush9525@fastmail.com>
Signed-off-by: Nelo-T. Wallus <n.wallus@sap.com>